Move Entries/Creation under Forms
